Creating a project and "Hero" component
Creating a project, adding libraries: "Tailwind", "TypeScript", "Router". Creating the "Hero" component, in which we created the "CustomButton" component. Also added styles to the file "tailwind.config.ts" and to the file "global.css". Also added a folder "types" in which created a file "index.ts" to describe the title, styles and function for the button. Added folder "folder" where all images for this project are placed.